config for google app engine (not working)
