+++ requirements.txt
@@ -1,0 +1,6 @@
+pdf2image==1.14.0
+pdfminer.six==20201018
+Flask==1.1.2
+fire==0.3.1
+Pillow==8.0.1
+gunicorn==20.0.4
